- Got radio buttons to look alright and function somewhat. - Still need to figure out the Worm part (WIP) - Plan out an attack for colour coding elements by their block
